i59: Move App.js under src
