+++ index.js
@@ -7,7 +7,7 @@ import { AppRegistry } from 'react-native';
 import store from 'rdx/store';
 import { Provider } from 'react-redux';
 import { SafeAreaProvider, initialWindowMetrics } from 'react-native-safe-area-context';
-import App from './App';
+import App from './src/App';
 import { name as appName } from './app.json';
 
 const AppContainer = () => (
